Refactor suggestion controller
