+++ suggestion/src/routes/suggestion.ts
@@ -1,25 +1,9 @@
 import * as express from 'express';
-import { suggestRecipes } from '../controller/suggestRecipe';
+import { suggestRecipeController } from '../controller/suggestRecipe';
 
 const router = express.Router();
 
 // POST /suggestion?ingredients=xyz
-router.post('/', (req, res) => {
-    // res.status(200).send('Test API endpoint');
-
-    const httpBody = req.body;
-    const numIngredients = httpBody.queryIngredients.length;
-
-    const testThreshold = 0.5;
-
-    const IDs = [];
-
-    for (let i = 0; i < numIngredients; i++) {
-        IDs.push(httpBody.queryIngredients[i].databaseID);
-    }
-
-    const recipeHashes = suggestRecipes(IDs, testThreshold);
-    res.status(200).send('Result: ' + recipeHashes);
-});
+router.post('/', suggestRecipeController);
 
 module.exports = router;
